feat(ButtonModal): Implement new modal header UI
