+++ src/components/Forms/AddServerForm/index.js
@@ -1,0 +1,4 @@
+import AddServerForm from './AddServerFormContainer';
+import AddServerFormHeader from './AddServerFormHeader';
+
+export { AddServerForm, AddServerFormHeader };
