+++ src/components/Modals/AddServerButtonModal/index.js
@@ -1,0 +1,3 @@
+import AddServerButtonModal from './AddServerButtonModal';
+
+export default AddServerButtonModal;
