+++ src/components/Modals/index.js
@@ -1,0 +1,4 @@
+import AddServerButtonModal from './AddServerButtonModal';
+import ButtonModal from './ButtonModal';
+
+export { AddServerButtonModal, ButtonModal };
