--- src/views/Home/AddServerForm/index.js
@@ -1,3 +1,0 @@
-import AddServerFormContainer from './AddServerFormContainer';
-
-export default AddServerFormContainer;
